requirements.txt
absl-py==0.11.0
anaconda-client==1.7.2
anaconda-navigator==1.10.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
atomicwrites==1.3.0
attrs==20.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.3
bleach==3.2.1
blis==0.4.1
bokeh==2.2.3
brotlipy==0.7.0
catalogue==1.0.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
clyent==1.2.2
conda==4.9.2
conda-build==3.20.5
conda-package-handling==1.7.2
conda-verify==3.4.2
cryptography==3.2.1
cycler==0.10.0
cymem==2.0.4
cytoolz==0.11.0
dask==2.30.0
datashape==0.5.4
decorator==4.4.2
defusedxml==0.6.0
distributed==2.30.1
docopt==0.6.2
en-core-web-sm==2.0.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
fsspec==0.8.3
future==0.18.2
gast==0.2.2
glob2==0.7
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
HeapDict==1.0.1
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab==2.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
libarchive-c==2.9
livelossplot==0.3.0
llvmlite==0.34.0
locket==0.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
msgpack==1.0.0
msgpack-numpy==0.4.3.2
multipledispatch==0.6.0
murmurhash==1.0.2
navigator-updater==0.2.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.4
numexpr==2.7.1
numpy==1.19.2
odo==0.5.1
olefile==0.46
opencv-python==4.5.1.48
openpyxl==2.5.14
opt-einsum==3.1.0
packaging==20.7
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.0
partd==1.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip==20.3
pipreqs==0.4.10
pkginfo==1.6.1
plac==0.9.6
preshed==3.0.2
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.20
pygame==2.0.1
Pygments==2.7.2
pyOpenSSL==20.0.0
pyparsing==2.2.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==5.3.1
pyzmq==20.0.0
QtAwesome==0.5.6
qtconsole==4.7.7
QtPy==1.9.0
regex==2018.1.10
requests==2.25.0
ruamel-yaml==0.15.87
scipy==1.5.2
Send2Trash==1.5.0
setuptools==50.3.2.post20201201
Shapely==1.7.1
six==1.11.0
sortedcontainers==2.3.0
soupsieve==2.0.1
spacy==2.3.2
SQLAlchemy==1.3.20
srsly==1.0.2
tables==3.6.1
tblib==1.7.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.1
toolz==0.11.1
tornado==6.1
tqdm==4.54.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.11
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wget==3.2
wheel==0.36.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
yarg==0.1.9
zict==2.0.0
zipp==3.4.0
